src/flick/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
boto3==1.13.16
botocore==1.13.20
cachetools==4.0.0
celery==4.4.2
certifi==2019.3.9
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==4.0.2
crayons==0.3.0
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-celery==3.3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-rest-framework==0.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
docutils==0.15.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
futures==3.1.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
goslate==1.5.1
httplib2==0.17.0
identify==1.4.11
idna==2.8
importlib-metadata==1.5.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
joblib==0.14.1
kombu==4.6.8
MarkupSafe==1.0
mysql-connector-python==8.0.20
mysqlclient==1.4.6
nltk==3.5
nodeenv==1.3.5
oauthlib==3.1.0
Pillow==7.1.2
pre-commit==2.1.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDictionary==1.5.2
python-crontab==2.5.1
python-dateutil==2.8.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3
redis==3.5.2
regex==2020.4.4
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
splinter==0.7.7
SQLAlchemy==1.3.1
sqlparse==0.3.1
toml==0.10.0
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.24.1
vine==1.3.0
virtualenv==16.7.9
webdriver-manager==2.3.0
Werkzeug==0.14.1
wordnet==0.0.1b2
zipp==3.0.0
